Simplify code with more hooks
